Move and capture generation

+++ src/bitboard.rs
@@ -31,11 +31,12 @@ impl BitBoard {
         (self.0 & square.as_bit()) != 0
     }
 
+    /// Returns the square index (0-63) of the least significant bit that is set
     /// Folding trick from chessprogramming.org
     /// https://www.chessprogramming.org/Matt_Taylor
-    pub fn next_bit(&mut self) -> u64 {
+    pub fn next_bit(&mut self) -> u8 {
         self.0 ^= self.0 - 1;
         let folded = self.0 ^ (self.0 >> 32);
-        LSB_64_TABLE[(folded as usize) * (0x78291ACF >> 26)] as u64
+        LSB_64_TABLE[(folded as usize) * (0x78291ACF >> 26)]
     }
 }
